configure cypress to run component test in react
