+++ cypress.config.ts
@@ -1,7 +1,15 @@
 import { defineConfig } from 'cypress';
+import viteConfig from './vite.config';
 export default defineConfig({
   e2e: {
     baseUrl: 'http://localhost:3000',
     supportFile: false,
+  },
+  component: {
+    devServer: {
+      framework: 'react',
+      bundler: 'vite',
+      viteConfig,
+    },
   },
 });
