+++ cypress/support/component.ts
@@ -1,0 +1,17 @@
+/* eslint-disable @typescript-eslint/no-namespace */
+/// <reference types="cypress" />
+import { mount } from 'cypress/react18';
+import './commands';
+
+Cypress.Commands.add('mount', (component, options) => {
+  return mount(component, options);
+});
+
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      searchWord(word: string): Chainable<void>;
+      mount: typeof mount;
+    }
+  }
+}
